% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gender_code_dictionary}
\alias{gender_code_dictionary}
\title{Gender code dictionary}
\format{
A data frame with three variables:
\describe{
\item{\code{entries}}{users' freetext responses}
\item{\code{broad}}{recoding returned by \code{recode_gender(method = "broad")}}
\item{\code{narrow}}{recoding returned by \code{recode_gender(method = "narrow")}}
}
}
\usage{
gender_code_dictionary
}
\description{
Inbuilt dictionary of potential freetext gender responses and ways to recode
them using \code{\link{recode_gender}}.
}
\keyword{datasets}
