% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendercodeR.R
\docType{package}
\name{gendercodeR}
\alias{gendercodeR}
\title{gendercodeR: A Package for Recoding Freetext Gender Data}
\description{
Provides built-in dictionaries and  a function \code{gender_recode}
to allow for easy automatic coding of common variations in free text
responses to the question \bold{"What is your gender?"}
}
