% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendercoder_app.R
\name{gendercoder_app}
\alias{gendercoder_app}
\title{Launch the gendercoder Shiny app}
\usage{
gendercoder_app(...)
}
\arguments{
\item{...}{arguments to pass to \code{shiny::runApp()}}
}
\value{
Called for its side effect of launching a Shiny app.
}
\description{
Code data interactively in a Shiny app that runs locally in RStudio or a web
browser using a \pkg{bs4Dash} interface. The app supports CSV, Stata, SPSS,
RDS, and R data files. Stata and SPSS files require the optional
\pkg{haven} package.
}
\examples{
if (interactive()) {
gendercoder_app()
}
}
