% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderCode.R
\name{genderRecode}
\alias{genderRecode}
\title{Gender column recoder}
\usage{
genderRecode(input, genderColName = "gender", method = "broad",
  outputColName = "gender_recode", saveUnrecognisedResponses = F)
}
\arguments{
\item{input}{A dataframe with a column with the name specified in gender column name}

\item{genderColName}{Gender information column text name}

\item{method}{explanation}

\item{outputColName}{Output column name}

\item{optionalReplacements}{margin parameters; vector of length 4 (see \code{\link[graphics]{par}})}
}
\value{
Returns the original dataframe
}
\description{
Recodes gender column to
}
\examples{


}
