% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderCode.R
\name{getCustomDictionary}
\alias{getCustomDictionary}
\title{of which will be treated as the values to be replaced and the second which will be treated as the
 replacement values (i.e., data.frame(c("cis  gender-male","trans-person" ), c("cis male", "transgender")))}
\usage{
getCustomDictionary(customDictionary = NA)
}
\value{
Returns the original dataframe with
}
\description{
of which will be treated as the values to be replaced and the second which will be treated as the
 replacement values (i.e., data.frame(c("cis  gender-male","trans-person" ), c("cis male", "transgender")))
}
