% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendercodeR.R
\docType{data}
\name{manylevels_en}
\alias{manylevels_en}
\title{manylevels_en}
\description{
<<<<<<< HEAD
A English dictionary for the recode_gender function with many levels
=======
A English dictionary for the recode_gender function that has many levels
>>>>>>> f379f8c41dc9dd327250cc91ba6f999d6a79cfca
}
