% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gender.R
\name{which_is_na}
\alias{which_is_na}
\title{which_is_na}
\usage{
which_is_na(x)
}
\arguments{
\item{x}{a list that contains responses and corresponding dictionary
matches #' responses without a match are stored with the NA or NULL
as the list reference.}
}
\value{
a vector of positions in input list where name is missing
}
\description{
finds terms without a dictionary match
}
\details{
Internal function to identify responses that are not
matched in the dictionary.
}
\examples{
x <- list("a"=1,"b" =2, "c" =3)
names(x)[c(1,3)] <-  NA
gendercoder:::which_is_na(x)
}
