% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendercodeR.R
\docType{_PACKAGE}
\name{gendercoder}
\alias{gendercoder}
\title{gendercoder: A Package for Recoding Freetext Gender Data}
\description{
Provides dictionaries and  a function \code{recode_gender}
to allow for easy automatic coding of common variations in free text
responses to the question \bold{"What is your gender?"}
}
