% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendercodeR.R
\docType{package}
\name{gendercoder}
\alias{gendercoder-package}
\alias{gendercoder}
\title{gendercoder: A Package for Recoding Freetext Gender Data}
\description{
Provides dictionaries and a function \code{recode_gender} to allow for easy automatic coding of common variations in free text responses to the question \strong{"What is your gender?"}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/gendercoder}
  \item Report bugs at \url{https://github.com/ropensci/gendercoder/issues}
}

}
\author{
\strong{Maintainer}: Yaoxiang Li \email{liyaoxiang@outlook.com} (\href{https://orcid.org/0000-0001-9200-1016}{ORCID})

Authors:
\itemize{
  \item Jennifer Beaudry \email{jbeaudry@swin.edu.au} (\href{https://orcid.org/0000-0003-1596-6708}{ORCID})
  \item Emily Kothe \email{emily.kothe@deakin.edu.au} (\href{https://orcid.org/0000-0003-1210-0554}{ORCID})
  \item Felix Singleton Thorn \email{fsingletonthorn@gmail.com} (\href{https://orcid.org/0000-0002-0237-6146}{ORCID})
  \item Rhydwyn McGuire \email{rhydwyn@rhydwyn.net}
  \item Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})
  \item Mathew Ling \email{mathewtyling@gmail.com} (\href{https://orcid.org/0000-0002-0940-2538}{ORCID})
}

Other contributors:
\itemize{
  \item Julia Silge (Julia reviewed the package (v. 0.0.0.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/435>) [reviewer]
  \item Elin Waring (Elin reviewed the package (v. 0.0.0.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/435>) [reviewer]
}

}
