% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendercoder_app.R
\name{gendercoder_app}
\alias{gendercoder_app}
\title{Launch Shiny App}
\usage{
gendercoder_app(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
Code your data interactively in a shiny app that runs locally in RStudio or your web browser (recommended). It does not connect to the web at all, so your data are completely private.
}
\examples{
\dontrun{ scivrs_app() }

}
