% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderCode.R
\name{genderRecode}
\alias{genderRecode}
\title{Gender column recoder}
\usage{
genderRecode(input, genderColName = "gender", method = "broad",
  outputColName = "gender_recode", missingValuesObjectName = NA,
  customDictionary = NULL)
}
\arguments{
\item{input}{A dataframe with a column with the name specified in gender column name}

\item{genderColName}{Gender information column text name}

\item{method}{"broad" or "narrow". Broad returns responses classified into "female", "male",
"androgynous", "non-binary", "nonbinary", "transgender", "transgender male", "transgender female",
 "intersex", "agender", "gender-queer". Narrow returns "female", "male", "cis female", "cis male", "other".
 @param customDictionary enter custom dictionary set up as a data frame with two columns, the first
 of which will be treated as the values to be replaced and the second which will be treated as the
 replacement values (i.e., data.frame(c("cis  gender-male","trans-person" ), c("cis male", "transgender")))}

\item{outputColName}{Output column name}

\item{missingValuesObjectName}{Provide object name for function to write unknown
responses and their locations to. If NA, the unrecognised responses will be
printed but not saved.}
}
\value{
Returns the original dataframe with
}
\description{
Recodes gender column to
}
\examples{


}
