% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_gender.R
\name{which_is_na}
\alias{which_is_na}
\title{finds terms without a dictionary match}
\usage{
which_is_na(x)
}
\arguments{
\item{x}{a list that contains responses and corresponding dictionary
matches #' responses without a match are stored with the NA or NULL
as the list reference.}
}
\value{

}
\description{
Internal function to identify responses that are not
matched in the dictionary.
}
\examples{

}
